% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_gtf.R
\name{create_tr_entry}
\alias{create_tr_entry}
\title{Create a new GTF entry for a novel terminal exon}
\usage{
create_tr_entry(start, end, pred_id, tr_ids, exon_copy, tran)
}
\arguments{
\item{start}{Integer scalar. Start of the novel exon.}

\item{end}{Integer scalar. End of the novel exon.}

\item{pred_id}{Integer scalar. ID of the predicted exon.}

\item{tr_ids}{Character vector. Transcript IDs of all transcripts that
contain the novel exon.}

\item{exon_copy}{GRanges object. All exons from the transcripts that contain
the novel exon.}

\item{tran}{GRanges object with all transcript annotations from the organism.}
}
\value{
GRanges object with the exon and transcript annotation entries of the
novel exon and all exons from the transcripts in tr_ids.
}
\description{
This function creates new GTF entries for a novel terminal exon and all
transcripts in which it is located. As input, the function takes a list of
transcript IDs and the GTF annotations of all exons from the transcripts. A
random exon from each transcript is copied and the start and end coordinates
are exchanged with those of the novel exon. The ranges of each transcript are
adjusted to include the novel exon. The \code{exon_number} and \code{exon_version} are
set to NA. The new \code{exon_id}, \code{transcript_id} and \code{transcript_name} are the
original values with a suffix that identifies the novel exon by its ID: e.g.
\code{exon_id}_ID or \code{transcript_name}_ID.
}
