% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DISCERNS.R
\docType{package}
\name{DISCERNS}
\alias{DISCERNS}
\title{DISCERNS}
\description{
The DISCERNS package predicts novel exons from RNA-seq data based on
the SJ.out.tab and the BAM file from STAR. The mapped splice-junctions are
compared with the genome annotation to discover novel exons. Existing genome
annotations in GTF format can be extended with the predicted exons.
}
\author{
Katharina M. Hembach
}
