% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_from_sj.R
\name{downstream_count}
\alias{downstream_count}
\title{Count mapped parts in a read from max-1 to 0 in the most downstream part
(from 5' to 3')}
\usage{
downstream_count(x)
}
\arguments{
\item{x}{GRangesList object with mapped parts of reads.}
}
\value{
data.frame with the range of each mapped part and the nr of each part
from max-1 to 0 (most downstream part in the read)
}
\description{
Count mapped parts in a read from max-1 to 0 in the most downstream part
(from 5' to 3')
}
