% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exondiscovery.R
\docType{package}
\name{exondiscovery}
\alias{exondiscovery}
\alias{exondiscovery-package}
\title{exondiscovery}
\description{
The exondiscovery package discovers novel exons from RNA-seq data based on
the SJ.out.tab and the BAM file from STAR. The mapped splice-junctions are
compared with the genome annotation to discover novel exons. Existing genome
annotations in GTF format can be extended with the predicted exons.
}
\author{
Katharina M. Hembach
}
