% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_reads.R
\name{filter_junction_reads}
\alias{filter_junction_reads}
\title{Filter reads with splice junctions from BAM file}
\usage{
filter_junction_reads(bam)
}
\arguments{
\item{bam}{The path to the BAM file.}
}
\value{
GAlignments object with junction reads
}
\description{
The function Filters all reads with "N" in the CIGAR string and that are
properly paired (-f 2) from a BAM file.
}
