% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_discovery_utils.R
\name{filter_trans_splice_junctions}
\alias{filter_trans_splice_junctions}
\title{Filter novel splice junctions based on overlapping genes}
\usage{
filter_trans_splice_junctions(sg, eg, sjg)
}
\arguments{
\item{sg}{Integer vector. Index of the genes that overlap with the SJ start.}

\item{eg}{Integer vector. Index of the genes that overlap with the SJ end.}

\item{sjg}{Integer vector. Index of the genes that overlap with the SJ.}
}
\value{
Logical scalar. FALSE if the start and end of the SJ overlap with
different genes.
}
\description{
This function filters splice junctions based on the genes that overlap with
the start and end of the splice junction, and that are crossed by the splice
junction.
}
