% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_discovery.R
\name{find_novel_exons}
\alias{find_novel_exons}
\title{Predict novel cassette exons from SJ.out.tab file from STAR}
\usage{
find_novel_exons(sj_filename, annotation, min_unique = 1,
  gzipped = FALSE, verbose = TRUE, bam)
}
\arguments{
\item{sj_filename}{Path to SJ.out.tab file.}

\item{annotation}{List with exon and intron annotation as GRanges. Created
with prepare_annotation().}

\item{min_unique}{Minimal number of reads required to map over a splice
junction.}

\item{gzipped}{A logical scalar. Is the input file gzipped? Default FALSE.}

\item{verbose}{Print messages about the progress.}

\item{bam}{Path to BAM file.}
}
\value{
Data frame with novel exons.
}
\description{
Predict novel cassette exons from SJ.out.tab file from STAR
}
