% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_novel_sj_reads.R
\name{get_reads}
\alias{get_reads}
\title{Import splice junction reads in a given genomic region}
\usage{
get_reads(bf, region)
}
\arguments{
\item{bf}{BamFile object.}

\item{region}{GRanges object specifying the genomic region from which to import
the reads.}
}
\value{
GAlignments object with all reads from \code{bf} overlapping \code{i}.
}
\description{
This function imports all splice junction reads within a specified genomic
region.
}
