% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_gtf.R
\name{get_transcripts}
\alias{get_transcripts}
\title{Find all transcripts that contain a given exon}
\usage{
get_transcripts(seqn, lend, start, end, rstart, strand, exons)
}
\arguments{
\item{seqn}{scalar, seqname of the novel exon}

\item{lend}{integer scalar, end of the upstream exon}

\item{start}{integer scalar, start of the novel exon}

\item{end}{integer scalar, end of the novel exon}

\item{rstart}{integer scalar, start of the downstream exon}

\item{strand}{factor scalar, strand of the novel exon (either "+" or "-")}

\item{exons}{GRanges object with exon annotations}
}
\value{
character vector with the transcript IDs of all complementary
  transcripts.
}
\description{
Given a novel exon, find the IDs of all transcripts that contain the up and
downstream exons of novel predicted exon.
}
