% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_from_sj.R
\name{identify_exon_end}
\alias{identify_exon_end}
\title{Determine the coordinates of a novel exon at the end of a splice junction}
\usage{
identify_exon_end(r, j_start, j_end, j_seqnames, j_strand)
}
\arguments{
\item{r}{GAlignments object with reads that contain the novel splice
junction.}

\item{j_start}{Integer scalar. Start of the splice junction.}

\item{j_end}{Integer scalar. End of the splice junction.}

\item{j_seqnames}{Factor scalar. seqname of the splice junction.}

\item{j_strand}{Factor scalar. Strand of the splice junction.}
}
\value{
data.frame with the coordinates of the predicted exon(s).
}
\description{
Given a set of reads that share a novel splice junction, determine the the
coordinates of a novel exon at the end of the splice junction. Only reads
with the splice junctions (the novel SJ and a downstream SJ) are considered.
If there are no reads with a second downstream SJ, take the range of the
longest mapped part as an approximation for the coordinates of the novel
exon (it probably is a terminal exon).
}
