% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_from_sj.R
\name{identify_exon_start}
\alias{identify_exon_start}
\title{Determine the coordinates of a novel exon at the start of a splice junction}
\usage{
identify_exon_start(r, j_start, j_end, j_seqnames, j_strand)
}
\arguments{
\item{r}{GAlignment object with reads that contain the novel splice
junction.}

\item{j_start}{Integer start of the splice junction.}

\item{j_end}{Integer end of the splice junction.}

\item{j_seqnames}{Factor seqname of the splice junction.}

\item{j_strand}{Factor strand of the splice junction.}
}
\value{
data.frame with the coordinates of the predicted exon(s)
}
\description{
Given a set of reads that share a novel splice junction, determine the the
coordinates of a novel exon at the start of the splice junction. Only reads
with the splice junctions (the novel SJ and an upstream SJ) are considered.
If there are no reads with a second upstream SJ, take the range of the
longest mapped part as an approximation for the coordinates of the novel
exon (it probably is a terminal exon).
}
