% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_novel_sj_reads.R
\name{import_novel_sj_reads_parallel}
\alias{import_novel_sj_reads_parallel}
\title{Import reads with novel splice junctions in parallel}
\usage{
import_novel_sj_reads_parallel(bam, yield_size = 2e+05, sj_unann,
  cores = 1)
}
\arguments{
\item{bam}{Character string. The path to the BAM file.}

\item{yield_size}{Integer scalar. Read the BAM file in chunks of this size.}

\item{sj_unann}{GRanges object with unannotated splice junctions.}

\item{cores}{Integer scalar. Number of cores to use. Default 1.}
}
\value{
GAlignments object with all reads from the BAM file that contain a
novel splice junctions.
}
\description{
This functions takes a BAM file and a GRanges object with unannotated splice
junctions as input and returns a GAlignments object with all reads from the
BAM file that contain any of the novel splice junctions. The BAM files is
read in parallel with mclapply using the specified number of \code{cores}.
}
