% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_discovery_utils.R
\name{match_sj_in_intron}
\alias{match_sj_in_intron}
\title{Match novel splice junctions within an intron}
\usage{
match_sj_in_intron(s, e)
}
\arguments{
\item{s}{Junctions from the start of the intron as GRanges.}

\item{e}{Junctions from the end of the intron as GRanges.}
}
\value{
A vector with the seqnames, end of the preceding exon, start of the
novel exon, end of the novel exon, start of the consecutive exon, and the
strand.
}
\description{
Match novel splice junctions within an intron
}
