% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_gtf.R
\name{new_left_terminal_transcript}
\alias{new_left_terminal_transcript}
\title{Create new GTF annotation for a novel left terminal exon}
\usage{
new_left_terminal_transcript(
  seqn,
  start,
  end,
  rstart,
  strand,
  pred_id,
  exons,
  tran
)
}
\arguments{
\item{seqn}{Character string or factor. seqname of the novel exon.}

\item{start}{Integer scalar. Start of the novel exon.}

\item{end}{Integer scalar. End of the novel exon.}

\item{rstart}{Integer scalar. Start of the downstream exon.}

\item{strand}{Character string or factor. Strand of the novel exon (either
"+" or "-").}

\item{pred_id}{Integer scalar. ID of the novel exon.}

\item{exons}{GRanges object. Exon annotations from a GTF file.}

\item{tran}{GRanges object. Transcript annotations from a GTF file.}
}
\value{
GRanges object with GTF annotations of the novel exon and all its
transcripts
}
\description{
Given a novel exon, the function identifies the IDs of all transcripts that
contain the downstream neighbouring exon (exon start is \code{rstart}) of the
novel exon. Each of the transcripts is copied and an entry for the the novel
exon is created. The \code{exon_number} and \code{exon_version} are set to NA. The new
\code{exon_id}, \code{transcript_id} and \code{transcript_name} are the original values with
a suffix that identifies the novel exon by its ID: e.g. \code{exon_id}_ID or
\code{transcript_name}_ID.
}
