% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_gtf.R
\name{new_transcript}
\alias{new_transcript}
\title{Create a new GTF entry for a novel exon}
\usage{
new_transcript(start, end, pred_id, tr_ids, exons)
}
\arguments{
\item{start}{Integer scalar. Start of the novel exon.}

\item{end}{Integer scalar. End of the novel exon.}

\item{pred_id}{Integer scalar. ID of the predicted exon.}

\item{tr_ids}{Character vector. Transcript IDs as returned by
\code{\link[=get_transcripts]{get_transcripts()}}.}

\item{exons}{GRanges object. Exon annotations from a GTF file.}
}
\value{
GRanges object with the exon annotation of the novel exon and all
exons from the transcripts in tr_ids.
}
\description{
Create GTF annotations for a novel exon, given a list of transcript IDs and
the GTF annotations of all exons. A random exon from each transcript is
copied and the start and end coordinates are exchanged with those of the
novel exon. The \code{exon_number} and \code{exon_version} are set to NA. The new
\code{exon_id}, \code{transcript_id} and \code{transcript_name} are the original values with
a suffix that identifies the novel exon by its ID: e.g. \code{exon_id}_ID or
\code{transcript_name}_ID.
}
