% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassette_exon.R
\name{predict_cassette_exon}
\alias{predict_cassette_exon}
\title{Predict novel cassette exons}
\usage{
predict_cassette_exon(sj_unann, introns)
}
\arguments{
\item{sj_unann}{GRanges object with unannotated splice junctions}

\item{introns}{GRanges object with intron annotations, e.g. the "introns" slot
from the prepare_annotation() return object.}
}
\value{
list with two slots: "ne" is a data.frame with the coordinates of
the identifies cassette exons. It has 6 columns: seqnames, lend, start, end,
rstart and strand. "sj" is a GRanges object with the SJs from sj_unann that
were not used to predict a cassette exons.
}
\description{
Novel cassette exons are predicted from pairs of novel splice junctions (SJ)
that are located within an annotated intron and share the start and end
coordinates of the intron:
X---------X   annotated intron
x---x         novel splice junction
      x---x   novel splice junction
X---NNN----X  predicted cassette exon (N)
}
\details{
First, the novel SJs are filtered: Only SJs that are located within an annotated
intron and that share their start or end coordinates with the intron are
retained. All introns that share both their start and end with a novel SJ are
tested for cassette exons. If the two novel SJs within an intron do not
overlap and are on the same strand, a novel cassette exon is predicted.
}
