% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_reads.R
\name{predict_jr_exon}
\alias{predict_jr_exon}
\title{Predict novel exons from reads with 2 junctions}
\usage{
predict_jr_exon(junc_reads, annotation)
}
\arguments{
\item{junc_reads}{GAlignments object with junction read.}

\item{annotation}{List with exon and intron annotation as GRanges. Created
with prepare_annotation().}
}
\value{
data.frame with the coordinates of the predicted novel exon. It has 6
columns: seqnames, lend, start, end, rstart and strand
}
\description{
Novel exons are predicted from novel combinations of annotated splice
junctions. Reads spanning two splice junctions are filtered from the set of
all spliced reads. Each of the splice junction pairs are compared with the
annotated splice junctions per transcript and already annotated splice
junction pairs are removed. In the last step, the predicted exon that is
defined by the two splice junctions is returned if it is contained within the
boundaries of any annotated gene. This prevents the prediction of false
positive exons from wrongly mapped reads.
}
