% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotation.R
\name{prepare_annotation}
\alias{prepare_annotation}
\title{Prepare exon and intron annotation from GTF file}
\usage{
prepare_annotation(GTF)
}
\arguments{
\item{GTF}{Path to GTF file}
}
\value{
List with exon and intron annotations as GRanges
}
\description{
Prepare exon and intron annotation from GTF file
}
