% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_discovery_utils.R
\name{transcript_range}
\alias{transcript_range}
\title{GRanges of a transcript}
\usage{
transcript_range(gr)
}
\arguments{
\item{gr}{GRanges object with all exons in a transcript}
}
\value{
GRange with start and end of the transcript
}
\description{
This function returns a GRange for a transcript given as a set of exons in a
GRanges object.
}
