% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_from_sj.R
\name{upstream_count}
\alias{upstream_count}
\title{Count mapped parts in a read from 1 to max (from 5' to 3')}
\usage{
upstream_count(x)
}
\arguments{
\item{x}{GRangesList object with mapped parts of reads.}
}
\value{
data.frame with the range of each mapped part and the nr of each part
from  1 to max (from 5' to 3').
}
\description{
Count mapped parts in a read from 1 to max (from 5' to 3')
}
