% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_discovery_utils.R
\name{which_exon_terminal}
\alias{which_exon_terminal}
\title{Determine terminal end of a splice junction}
\usage{
which_exon_terminal(j, txdb = txdb, gtxdb = gtxdb, ebyTr = ebyTr)
}
\arguments{
\item{j}{GRanges object. Novel splice junction.}

\item{txdb}{TxDb object, e.g. the "txdb" slot from the \code{\link[=prepare_annotation]{prepare_annotation()}}
return object.}

\item{gtxdb}{GRanges object. All genes from the txdb parameter, e.g. obtained
with \code{GenomicFeatures::genes(txdb)}.}

\item{ebyTr}{GRangesList object. All exons per transcript of the txdb
parameter, e.g. obtained with \code{GenomicFeatures::exonsBy(txdb, by = "tx", use.names = TRUE)}.}
}
\value{
Character string; either "start", "end" or "NA".
}
\description{
This function computes if a splice junction (SJ) splices to a terminal exons.
If the exon at the start of the SJ is terminal, the function returns "start",
"end" if it is the exon at the end and "NA" if none of the exons are
terminal.
}
\section{Details}{

The function first overlaps the SJ with all genes and then with all
transcripts of the overlapping gene.
}

