% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_gtf.R
\name{extend_gtf}
\alias{extend_gtf}
\title{Extend GTF file with predicted exons}
\usage{
extend_gtf(gtf, pred)
}
\arguments{
\item{gtf}{GTF annotations, either the path to the GTF file or a GRanges
object}

\item{pred}{data.frame with predicted novel exons as returned by
find_novel_exons()}
}
\value{
GRanges object with annotations from the GTF file and extended with
  the novel exons
}
\description{
Add predicted novel exons to the correct transcripts in a GTF annotation.
}
