% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_reads.R
\name{filter_junction_reads}
\alias{filter_junction_reads}
\title{Junction reads from BAM file}
\usage{
filter_junction_reads(bam, yield_size = 2e+05, lib_type = "PE",
  stranded = "reverse")
}
\arguments{
\item{bam}{The path to the BAM file.}

\item{yield_size}{Integer scalar. The number of reads that should be read in
each chunk.}

\item{lib_type}{Character scalar. Type of the sequencing library: either "SE"
(single-end) or "PE" (paired-end). Default: "PE"}

\item{stranded}{Character scalar. Strand type of the sequencing protocol:
"unstranded" for unstranded protocols; "forward" or "reverse" for stranded
protocols. "forward" refers to protocols where the first read comes from the
forward (sense) strand. Default: "reverse"}
}
\value{
GAlignments object with all junction reads from the bam file.
}
\description{
The function takes a BAM file as input and returns all reads with "N" in the
CIGAR string and that are properly paired.
}
\details{
The yieldSize param determines the runtime: The bigger, the faster. If possible, use at
least 200000.
}
