% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_reads.R
\name{filter_junction_reads}
\alias{filter_junction_reads}
\title{Junction reads from BAM file}
\usage{
filter_junction_reads(bam, yield_size = 2e+05, lib_type = "PE",
  stranded = "reverse")
}
\arguments{
\item{bam}{Character string. The path to the BAM file.}

\item{yield_size}{Integer scalar. Read the BAM file in chunks of this size.}

\item{lib_type}{Character string. Type of the sequencing library: either "SE"
(single-end) or "PE" (paired-end). Default "PE".}

\item{stranded}{Character string. Strand type of the sequencing protocol:
"unstranded" for unstranded protocols; "forward" or "reverse" for stranded
protocols. In a "forward" protocol, the first read in a pair (or sinlge-end
reads) comes from the forward (sense) strand and in a "reverse" protocol
from the reverse (antisense) strand. See the \href{https://salmon.readthedocs.io/en/latest/library_type.html#fraglibtype}{Salmon documentation}
for an explanation of the different fragment library types. Default
"reverse".}
}
\value{
GAlignments object with all junction reads from the bam file.
}
\description{
The function takes a BAM file as input and returns all reads with "N" in the
CIGAR string and that are properly paired.
}
\details{
The yieldSize param determines the runtime: The bigger, the faster. If possible, use at
least 200000.
}
