% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_discovery.R
\name{find_novel_exons}
\alias{find_novel_exons}
\title{Novel exon prediction}
\usage{
find_novel_exons(sj_filename, annotation, min_unique = 1,
  gzipped = FALSE, verbose = TRUE, bam, single_sj = TRUE,
  read_based = TRUE, yieldSize = 2e+05)
}
\arguments{
\item{sj_filename}{Path to SJ.out.tab file.}

\item{annotation}{List with exon and intron annotation as GRanges. Created
with prepare_annotation().}

\item{min_unique}{Minimal number of reads required to map over a splice
junction.}

\item{gzipped}{A logical scalar. Is the input file gzipped? Default FALSE.}

\item{verbose}{Print messages about the progress.}

\item{bam}{Path to BAM file.}

\item{single_sj}{A logical scalar. Should single novel splice junctions be
used for prediction? Requires a BAM file.}

\item{read_based}{A logical scalar. Should novel exons be predicted from
reads or read-pairs with two novel splice junctions? Requires a BAM file.}

\item{yieldSize}{Integer scalar. Read the BAM file in chunks of this size.}
}
\value{
Data frame with novel exons.
}
\description{
Predict novel exons based on the SJ.out.tab file from STAR and/or a BAM file.
}
\details{
The function predicts different types of exons. The different prediction
modes require different inputs.
}
