% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_gtf.R
\name{get_transcripts}
\alias{get_transcripts}
\title{Find all transcripts that contain a given exon}
\usage{
get_transcripts(seqn, lend, start, end, rstart, strand, exons)
}
\arguments{
\item{seqn}{Scalar, seqname of the novel exon.}

\item{lend}{Integer scalar, end of the upstream exon.}

\item{start}{Integer scalar, start of the novel exon.}

\item{end}{Integer scalar, end of the novel exon.}

\item{rstart}{Integer scalar, start of the downstream exon.}

\item{strand}{Factor scalar, strand of the novel exon (either "+" or "-").}

\item{exons}{GRanges object with exon annotations.}
}
\value{
Character vector with transcript IDs.
}
\description{
Given a novel exon, find the IDs of all transcripts that contain the upstream
and downstream exons.
}
