% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_from_sj.R
\name{identify_exon_from_sj}
\alias{identify_exon_from_sj}
\title{Identify a novel exon from a single novel splice junction}
\usage{
identify_exon_from_sj(df, reads, txdb, cores)
}
\arguments{
\item{df}{data.frame with novel splice junctions. Following columns are
required: \code{seqnames}, \code{start}, \code{end}, \code{strand} and \code{touching}.}

\item{reads}{GAlignments object. Reads that contain novel splice
junctions, e.g. obtained with import_novel_sj_reads().}

\item{txdb}{TxDb object, e.g. the "txdb" slot from the \code{\link[=prepare_annotation]{prepare_annotation()}}
return object.}

\item{cores}{Integer scalar. Number of cores to use. Default 1.}
}
\value{
data.frame with the coordinates of the identified novel exons from all
splice junctions in df. It has 6 columns:
\code{seqnames}, \code{lend}, \code{start}, \code{end}, \code{rstart} and \code{strand}.
}
\description{
Determine the size of the novel exon (unpaired internal splice junctions (SJ)
and terminal junctions) based on the touching anntated exons and the RNA-seq
reads. If a novel SJ touches either the start or end of an annotated exon and
we have supporting reads with two SJs (the novel SJ and an annotated one)
then we know the size of the novel exon and the coordinates of the
neighbouring exons. If a novel SJ touches annotated exons on both ends, then
the novel exon is either terminal or it overlaps with existing exons. In case
of a terminal exon, the function returns NA for the location of the
neighbouring exon.
}
