% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_gtf.R
\name{new_transcript}
\alias{new_transcript}
\title{Create a new GTF entry for a novel exon}
\usage{
new_transcript(start, end, tr_ids, exons)
}
\arguments{
\item{start}{Integer scalar, start of the novel exon.}

\item{end}{Integer scalar, end of the novel exon.}

\item{tr_ids}{Character vector with transcrip ids as returned by
\code{\link[=get_transcripts]{get_transcripts()}}}

\item{exons}{GRanges object with exon annotations from a GTF file.}
}
\value{
GRanges object with the exon annotation of the novel exon and all
exons from the transcripts in tr_ids.
}
\description{
Create GTF annotations for a novel exon, given a list of transcript IDs and
the GTF annotations of all exons. A random exon from each transcript is
copied and the start and end coordinates are exchanged with those of the
novel exon. The exon number and exon version are set to NA and the new exon
ID is the original exon id with a suffix that identifies the novel exon:
ID_me/exon_start:end
}
