% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_reads.R
\name{predict_jrp_exon}
\alias{predict_jrp_exon}
\title{Predict novel exons from read pairs with two splice junctions}
\usage{
predict_jrp_exon(
  junc_reads,
  annotation,
  read_length = 101,
  overhang_min = 12,
  min_intron_size = 21
)
}
\arguments{
\item{junc_reads}{GAlignments object with junction reads.}

\item{annotation}{List with exon and intron annotation as GRanges. Created
with \code{\link[=prepare_annotation]{prepare_annotation()}}.}

\item{read_length}{Integer scalar. Length of your reads in bps. Default 101.}

\item{overhang_min}{Integer scalar. Minimum overhang length for splice
junctions on both sides as defined by the \code{--outSJfilterOverhangMin}
parameter of STAR. Use the minimum of the values for canonical splice junctions
(value (2) to (4)). You do not have to set this parameter if you used the
default values from STAR. Default 12.}

\item{min_intron_size}{Integer scalar. Minimum intron size
(\code{--alignIntronMin} parameter of STAR). You do not have to set this
parameter if you used the default values from STAR. Default 21.}
}
\value{
data.frame with the coordinates of the predicted novel exon. It has 6
columns: \code{seqnames}, \code{lend}, \code{start}, \code{end}, \code{rstart} and \code{strand}.
}
\description{
Novel exons are predicted from paired-end reads where each read spans one
splice junction. First, the read pairs are filtered and the distance between
the end of the first junction and the start of the second junction has to be
\verb{< 2 * (read_length - overhang_min) + min_intron_size}. Here, \code{read_length}
is the length of the reads, \code{overhang_min} is the minmal required read
overhang over a splice junction of the alignment tool and \code{min_intron_size}
is the minimal required intron length of the alignment tool. For example,
paired-end reads with a lenght of 101 nts and a minimal overhang of 6 and a
minimal intron length of 21 allow a distance of at most 211 nucleotides
between the two splice junctions: \code{2 * (101 - 6) + 21 = 211}. If the distance
between the two splice junctions exceeds the limit, it cannot be guaranteed
that the junctions are connected to the same exon. Splice junction pairs that
are already annotated in a transcript are removed. Novel exons are predicted
from the remaining splice junction pairs.
}
