% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annotation.R
\name{prepare_annotation}
\alias{prepare_annotation}
\title{Prepare anntation}
\usage{
prepare_annotation(gtf_file)
}
\arguments{
\item{gtf_file}{Character string. Path to a GTF file.}
}
\value{
List with exon and intron annotations as GRanges and the \code{TxDB}
object. The slots are \code{exons}, \code{introns} and \code{txdb}.
}
\description{
Read exon and intron annotation from a GTF file.
}
\examples{
gtf <- system.file("extdata", "selected.gtf", package = "exondiscovery", 
                   mustWork = TRUE)
anno <- prepare_annotation(gtf)
names(anno)
}
