% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_discovery_utils.R
\name{sj_touching_exon}
\alias{sj_touching_exon}
\title{Determine which side of the splice junction touches an exon}
\usage{
sj_touching_exon(seqnames, start, end, strand, exons)
}
\arguments{
\item{seqnames}{String: splice junction seqname.}

\item{start}{Integer: start of splice junction.}

\item{end}{Integer: end of splice junction.}

\item{strand}{Factor: strand of splice junction ("-" or "+").}

\item{exons}{Exon annotation as GRanges.}
}
\value{
A string that defines which side of the splice junction is touching
an exon: "both", "start", "end", NA if the splice junction does not touch
an exon or if the exons come from different genes.
}
\description{
Find all exons with the same strand as the splice junction and that touch it.
If a sj touches two exons, the exons must come from the same gene.
}
