% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_discovery_utils.R
\name{transcript_range}
\alias{transcript_range}
\title{Transcript range}
\usage{
transcript_range(gr)
}
\arguments{
\item{gr}{GRanges object. All exons of a transcript.}
}
\value{
GRanges object with the start and end of the transcript.
}
\description{
This function returns a GRanges object with the start and end of a transcript
and takes a GRanges object with the transcript's exons as input.
}
