% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exon_discovery_utils.R
\name{which_exon_terminal}
\alias{which_exon_terminal}
\title{Determine terminal end of a splice junction}
\usage{
which_exon_terminal(j, txdb = txdb, gtxdb = gtxdb, ebyTr = ebyTr)
}
\arguments{
\item{j}{Novel splice junction as GRange.}

\item{txdb}{TxDb object, e.g. the txdb slot from the prepare_annotation()
return object}

\item{gtxdb}{All genes from the TxDB object}

\item{ebyTr}{All exons from the TxDB object per transcript.}
}
\value{
Scalar with either "start", "end" or "NA"
}
\description{
This function computes if a splice junction (SJ) splices to a terminal exons.
If the exon at the start of the SJ is terminal, the function returns "start",
"end" if it is the exon at the end and "NA" if none of the exons are terminal
}
